refactor(desktop): organize workspace git publish hooks
